Show correct stat ranks for players with 0 chars
Previously fame rank, exp rank, and account fame rank were determined
based on whether a particular regex returned the full string (???) or
if the value didn't contain a set of parentheses. Players with 0 chars
had a ranking, though it was a very low ranking.

Now, due to some site changes along the line, the way the ranks display
is different - players with no characters (or hidden characters) don't
even have a ranking, and simply display 0 fame and exp, along with
whatever account fame they have, and players with any other number of
characters will display their rank.

As such, the only determining factor for the 'null' rank, as it were,
should be whether the string contains a paren. If so, a rank is
present. Otherwise, no rank is present, and -1 should be returned.

Fixes #17