Ensure API version header is in proper format
Because of the nature of git's smudge-clean filter being maintained
outside of the code, we can't depend on '$Id$' being replaced
successfully if it is improperly configured. As such, a check is
performed against the "commit hash" to ensure it is actually in the
correct format. (case-insensitive letters a-f and digits only)