Improve deploy script security, filtering
The deployment script previously allowed any user with access to the
target server's web domain to run it, and would attempt to run after
any push to the repository. After these changes, only one branch
is tracked (in this case, master), and attempts at access are
restricted to whitelisted IPs and GitHub.

Fixes #28