Null equipment hotfix
Due to a change in RealmEye's HTML, the "data-vars" attributes of
character equipment were removed from the DOM, breaking our parsing of
character equipment.

To fix this, we reference the HREF of the item links
(eg. /wiki/demon-blade) and compare the string against the full items
array. This is unfortunately rather slow, but unavoidable.

In the process, item assignment has been cleaned up to be more
data-driven.

P.S. Also corrected indentation for the column processing section.