Make RealmEyeAPIUtils properties static
Because this class will be loaded in other classes in the future to get
active configuration values and an application logger object, its
properties need to be consistent when accessed from different sources.
This has the added benefit of reducing toil when the config and logger
properties are already initialized.